Added support for capistrano deployments
Allowed $jardest to be passed in. Now if cap_deploy is true in caller
node puppet won't do anything but set the sakaioae.ar symlink dest to
what's specified.

There is an accompanying pull request for the main puppet-oae project
that depend on this, so I'll wait to submit until this is merged. BTW I
tested this on qa.academic.rsmart.com and all is well for either case
(cap deploy or not)